+++ src/commands/servlist.cpp
@@ -1,0 +1,8 @@
+#include "common.h"
+#include "servlist.h"
+
+void
+cmd_servlist(const char *data)
+{
+    (void) data;
+}