+++ src/commands/servlist.h
@@ -1,0 +1,8 @@
+#ifndef CMD_SERVLIST_H
+#define CMD_SERVLIST_H
+
+__SWIRC_BEGIN_DECLS
+void cmd_servlist(const char *);
+__SWIRC_END_DECLS
+
+#endif