+++ src/include/commandhelp.h
@@ -474,6 +474,16 @@ static const char *say_usage[] = {
     "",
 };
 
+static const char *servlist_usage[] = {
+    "usage: /servlist [<mask> [<type>]]",
+    "",
+    "This command is used to list services currently connected to the network",
+    "and visible to the user issuing the command. The optional parameters may",
+    "be used to restrict the result of the query (to matching services names,",
+    "and services type).",
+    "",
+};
+
 static const char *set_usage[] = {
     "usage: /set [[setting] [value]]",
     "",